refactor: split tenant lifecycle services
